server/socratic_service/requirements.txt
flask
flask-cors
chromadb
google-generativeai
pypdf
python-dotenv
werkzeug
